new file for calculating OF scale factor
