+++ analysis/computeZllOFscaleFactor.cpp
@@ -1,0 +1,1154 @@
+#include "TFile.h"
+#include "TMath.h"
+#include "TF1.h"
+#include "TProfile.h"
+#include "TLegend.h"
+#include "TCanvas.h"
+#include "TStyle.h"
+#include "TTree.h"
+#include "TH2D.h"
+#include "TH1D.h"
+#include "TGaxis.h"
+#include "THStack.h"
+#include "TLorentzVector.h"
+#include "TRandom3.h"
+#include "TGraphErrors.h"
+
+#include "interface/MT2Sample.h"
+#include "interface/MT2Region.h"
+#include "interface/MT2Analysis.h"
+#include "interface/MT2Estimate.h"
+#include "interface/MT2EstimateTree.h"
+#include "interface/MT2Config.h"
+
+
+#include "../interface/MT2DrawTools.h"
+
+
+#include <iostream>
+#include "string.h"
+#include <fstream>
+
+
+#define mt2_cxx
+#include "interface/mt2.h"
+
+
+bool shapeNorm = false;
+bool HFveto = false;
+
+double lumiErr = 0.12;
+
+void randomizePoisson( TH1* histo );
+
+
+void drawYields( MT2Config cfg, MT2Analysis<MT2EstimateTree>* data, std::vector<MT2Analysis<MT2EstimateTree>* >  bgYields, const std::string& saveName, const std::string& varName, const std::string& selection, int nBins, float xMin, float xMax, std::string axisName="", const std::string& units="" );
+
+void drawYieldsScaled( MT2Config cfg, MT2Analysis<MT2EstimateTree>* data, std::vector<MT2Analysis<MT2EstimateTree>* >  bgYields, const std::string& saveName, const std::string& varName, const std::string& selection, int nBins, float xMin, float xMax, std::string axisName="", const std::string& units="", float scale =1 );
+
+
+float computeScaleFactor( MT2Config cfg, MT2Analysis<MT2EstimateTree>* data, std::vector<MT2Analysis<MT2EstimateTree>* >  bgYields, const std::string& saveName, const std::string& varName, const std::string& selection, int nBins, float xMin, float xMax, std::string axisName="", const std::string& units="" );
+
+
+
+int main(int argc, char* argv[]){
+
+  std::string regionsSet = "zurich";
+
+
+  if( argc>2 ) {
+    std::string normType(argv[2]);
+    if( normType=="lumi" ) shapeNorm=false;
+    else if( normType=="shape" ) shapeNorm=true;
+    else {
+      std::cout << "-> Only 'lumi' and 'shape' are supported normTypes." << std::endl;
+      exit(17);
+    }
+  }
+
+
+  if( argc<2 ) {
+    std::cout << "USAGE: ./coputeZllGammaRatio [configFileName] regionSet" << std::endl;
+    std::cout << "Exiting." << std::endl;
+    exit(11);
+  }
+
+  std::string configFileName(argv[1]);
+  MT2Config cfg( configFileName);
+  std::string samplesFileName = "../samples/samples_" + cfg.mcSamples() + ".dat"; 
+  std::string samples = cfg.mcSamples();
+
+  regionsSet = cfg.regionsSet();
+
+
+  std::string outputdir = cfg.getEventYieldDir() + "/zllPurity";
+  std::string outputdir_of = cfg.getEventYieldDir() + "/zllPurity_of";
+
+  // std::string outputdir( Form("ZllData_%s", configFileName.c_str() ) );
+  // std::string outputdir_of( Form("ZllData_OF_%s", configFileName.c_str()  ) );
+
+
+  std::cout << "-> Using regions: " << regionsSet << std::endl;
+
+  TH1::AddDirectory(kFALSE); // stupid ROOT memory allocation needs this
+
+  double intpart;
+  double fracpart = modf(cfg.lumi(), &intpart);
+  std::string suffix;
+  if( fracpart>0. )
+    suffix = std::string( Form("_%.0fp%.0ffb", intpart, 10.*fracpart ) );
+  else
+    suffix = std::string( Form("_%.0ffb", intpart ) );
+  //outputdir += suffix;
+  // outputdir_of += suffix;
+  
+  system(Form("mkdir -p %s", outputdir.c_str()));
+
+
+
+
+  std::string ZllDir = cfg.getEventYieldDir() + "/zllControlRegion";
+  std::string ZllDir_of = cfg.getEventYieldDir() + "/zllControlRegion";
+
+  //  std::string ZllDir = "ZllPurity_" + configFileName;
+  //  std::string ZllDir_of = "ZllPurity_OF_" + configFileName;
+
+
+  MT2Analysis<MT2EstimateTree>* Zll = MT2Analysis<MT2EstimateTree>::readFromFile(Form("%s/ZllPurityTrees.root", ZllDir.c_str() ), "DYJets");
+  if( Zll==0 ) {
+    std::cout << "-> Please run zllPurityTrees first. I need to get the yields from there." << std::endl;    std::cout << "-> Thank you for your cooperation." << std::endl;    exit(197);
+  }
+
+
+
+  
+  MT2Analysis<MT2EstimateTree>* qcd = MT2Analysis<MT2EstimateTree>::readFromFile(Form("%s/ZllPurityTrees.root", ZllDir.c_str()  ), "QCD");
+  MT2Analysis<MT2EstimateTree>* top = MT2Analysis<MT2EstimateTree>::readFromFile(Form("%s/ZllPurityTrees.root", ZllDir.c_str() ), "Top");
+  MT2Analysis<MT2EstimateTree>* wjets = MT2Analysis<MT2EstimateTree>::readFromFile(Form("%s/ZllPurityTrees.root", ZllDir.c_str() ), "WJets");
+  MT2Analysis<MT2EstimateTree>* zjets = MT2Analysis<MT2EstimateTree>::readFromFile(Form("%s/ZllPurityTrees.root", ZllDir.c_str() ), "ZJets");
+
+  
+ 
+ //OPPOSITE FLAVOR TREES
+  MT2Analysis<MT2EstimateTree>* Zll_of = MT2Analysis<MT2EstimateTree>::readFromFile(Form("%s/ZllPurityTrees_of.root", ZllDir_of.c_str() ), "DYJets");
+
+  MT2Analysis<MT2EstimateTree>* qcd_of = MT2Analysis<MT2EstimateTree>::readFromFile(Form("%s/ZllPurityTrees_of.root", ZllDir_of.c_str()  ), "QCD");
+  MT2Analysis<MT2EstimateTree>* top_of = MT2Analysis<MT2EstimateTree>::readFromFile(Form("%s/ZllPurityTrees_of.root", ZllDir_of.c_str() ), "Top");
+  MT2Analysis<MT2EstimateTree>* wjets_of = MT2Analysis<MT2EstimateTree>::readFromFile(Form("%s/ZllPurityTrees_of.root", ZllDir_of.c_str() ), "WJets");
+  MT2Analysis<MT2EstimateTree>* zjets_of = MT2Analysis<MT2EstimateTree>::readFromFile(Form("%s/ZllPurityTrees_of.root", ZllDir_of.c_str() ), "ZJets");
+  
+  MT2Analysis<MT2EstimateTree>* data_of = MT2Analysis<MT2EstimateTree>::readFromFile(Form("%s/data_of.root", ZllDir_of.c_str() ) , "data_of");
+
+  Zll_of->setFullName("Z+jets");
+  wjets_of->setFullName("W+jets");
+  zjets_of->setFullName("Z#nu#nu+jets");
+  data_of->setFullName("Data");
+ 
+
+  MT2Analysis<MT2EstimateTree>* data = MT2Analysis<MT2EstimateTree>::readFromFile(Form("%s/data.root", ZllDir.c_str() ) , "data");
+ 
+
+  data->setFullName("Data");
+
+  Zll->setFullName("Z+jets");
+  wjets->setFullName("W+jets");
+  zjets->setFullName("Z#nu#nu+jets");
+
+
+
+  std::vector<MT2Analysis<MT2EstimateTree>* > bgYields; 
+  bgYields.push_back( Zll );
+  bgYields.push_back( qcd );
+  bgYields.push_back( wjets );
+  bgYields.push_back( zjets );
+  bgYields.push_back( top );
+
+
+  // drawMll( outputdir, bgYields, data,  0 , cfg.lumi() );
+ 
+  
+  std::vector<MT2Analysis<MT2EstimateTree>* > bgYields_of; 
+  bgYields_of.push_back( Zll_of );
+  bgYields_of.push_back( qcd_of );
+  bgYields_of.push_back( wjets_of );
+  bgYields_of.push_back( zjets_of );
+  bgYields_of.push_back( top_of );
+  
+
+  //drawMll( outputdir_of, bgYields_of, data_of,  1, cfg.lumi() );
+  
+
+
+
+
+
+
+
+
+  MT2DrawTools::setStyle();
+ 
+
+
+  std::string    selection = "weight*(abs(Z_mass-91.19)<20 && Z_pt>0 && nJetHF30==0)";
+  std::string selection_mass = "weight*(Z_pt>0 && nJetHF30==0 )";
+  
+  /*
+  std::string selection_mass = "weight*(Z_mass>50 && Z_pt>180)";
+  std::string      selection_mass_el = "weight*(Z_mass>50 && Z_pt>180 && Z_lepId==11)";
+  std::string      selection_mass_mu = "weight*(Z_mass>50 && Z_pt>180 && Z_lepId==13)";
+
+
+  std::string      selection_mass = "weight";
+ 
+  std::string      selection =  "weight*(Z_mass>20 && Z_pt> 0 )";
+  //selection2 = "weight*(Z_mass>80)";
+ 
+  std::string      selection3 = "weight*(Z_mass>70)";
+  std::string      selection4 ="weight*(Z_mass>80)";
+
+  std::string selection = "weight*(Z_pt>180 && abs(Z_mass-91.19)<20)";
+  */
+  drawYields( cfg, data, bgYields, "Z_mass" , "Z_mass" , selection_mass, 50, 50, 150, "M_{ll}", "GeV" );
+
+
+
+  std::string    selection_of = "weight*(abs(Z_mass)>20 && Z_pt>0 && nJetHF30==0)";
+  // drawYields( cfg, data_of, bgYields_of, "Z_mass" , "Z_mass" , selection_of, 20, 0, 200, "M_{ll}", "GeV" );
+
+ 
+
+  float scaleFactor =  computeScaleFactor( cfg, data_of, bgYields_of, "Z_mass_of" , "Z_mass" , selection_of, 20, 20, 200, "M_{ll}", "GeV" );
+
+  drawYieldsScaled( cfg, data_of, bgYields_of, "Z_mass_of_scaled" , "Z_mass" , selection_of, 20, 20, 200, "M_{ll}", "GeV" , scaleFactor);
+  drawYieldsScaled( cfg, data, bgYields, "Z_mass_scaled" , "Z_mass" , selection_mass, 50, 50, 150, "M_{ll}", "GeV" , scaleFactor);
+
+
+  std::ofstream ofs ("scaleFactorOF.txt", std::ofstream::out);
+  ofs<< scaleFactor;
+  ofs.close();
+
+
+  return 0;
+}
+
+
+
+
+
+
+
+
+
+
+
+
+
+void drawYields( MT2Config cfg, MT2Analysis<MT2EstimateTree>* data, std::vector<MT2Analysis<MT2EstimateTree>* >  bgYields, const std::string& saveName, const std::string& varName, const std::string& selection, int nBins, float xMin, float xMax, std::string axisName, const std::string& units ) {
+
+
+  float binWidth = (xMax-xMin)/nBins;
+  if( axisName=="" ) axisName = varName;
+
+
+
+  std::vector<int> colors;
+  if( bgYields.size()==3 ) { // estimates
+    colors.push_back(402); 
+    colors.push_back(430); 
+    colors.push_back(418); 
+  } else { // mc
+    colors.push_back(430); // other=zll
+    colors.push_back(401); // qcd
+    colors.push_back(417); // w+jets
+    colors.push_back(419); // z+jets
+    colors.push_back(855); // top
+    //colors.push_back(); // other
+  }
+
+  std::string fullPathPlots = cfg.getEventYieldDir() + "/plotsDataMC";
+  system( Form("mkdir -p %s", fullPathPlots.c_str()) );
+
+  TH1::AddDirectory(kTRUE); // stupid ROOT memory allocation needs this
+
+  std::set<MT2Region> MT2Regions = data->getRegions();
+  
+  for( std::set<MT2Region>::iterator iMT2 = MT2Regions.begin(); iMT2!=MT2Regions.end(); ++iMT2 ) {
+  
+    MT2Region thisRegion( (*iMT2) );
+
+    TTree* tree_data = data->get(thisRegion)->tree;
+    TH1D* h1_data = new TH1D("h1_data", "", nBins, xMin, xMax );
+    tree_data->Project( "h1_data", varName.c_str(), selection.c_str() );
+
+    TGraphAsymmErrors* gr_data = MT2DrawTools::getPoissonGraph(h1_data);
+    gr_data->SetMarkerStyle(20);
+    gr_data->SetMarkerSize(1.2);
+
+
+    std::vector< TH1D* > histos_mc;
+    for( unsigned i=0; i<bgYields.size(); ++i ) { 
+      TTree* tree_mc = (bgYields[i]->get(thisRegion)->tree);
+      std::string thisName = "h1_" + bgYields[i]->getName();
+      TH1D* h1_mc = new TH1D( thisName.c_str(), "", nBins, xMin, xMax );
+      h1_mc->Sumw2();
+      if( selection!="" )
+	//tree_mc->Project( thisName.c_str(), varName.c_str(), Form("%s/puWeight", selection.c_str()) );
+	tree_mc->Project( thisName.c_str(), varName.c_str(), Form("%s", selection.c_str()) );
+      else
+        tree_mc->Project( thisName.c_str(), varName.c_str(), "" );
+      histos_mc.push_back(h1_mc);
+    }
+
+    TH1::AddDirectory(kFALSE); // stupid ROOT memory allocation needs this
+
+    TH1D* mc_sum;
+    for( unsigned i=0; i<histos_mc.size(); ++i ) { 
+      if( i==0 ) {
+        mc_sum = new TH1D( *histos_mc[i] );
+        mc_sum->SetName("mc_sum");
+      } else {
+        mc_sum->Add( histos_mc[i] );
+      }
+    }
+
+    std::cout << "Integrals: " << h1_data->Integral(0, nBins+1) << "\t" << mc_sum->Integral(0, nBins+1) << std::endl;
+    float scaleFactor = h1_data->Integral(0, nBins+1)/mc_sum->Integral(0, nBins+1);   
+    if( shapeNorm ) 
+      std::cout << "SF: " << scaleFactor << std::endl;
+
+    double error_data;
+    double integral_data = h1_data->IntegralAndError(0, nBins+1, error_data);
+
+    double error_mc;
+    double integral_mc = mc_sum->IntegralAndError(0, nBins+1, error_mc);
+
+    double error_datamc = MT2DrawTools::getSFError(integral_data, error_data, integral_mc, error_mc);
+
+    TH1D* histo_mc;
+    THStack bgStack("bgStack", "");
+    for( unsigned i=0; i<histos_mc.size(); ++i ) { 
+      int index = bgYields.size() - i - 1;
+      histos_mc[index]->SetFillColor( colors[index] );
+      histos_mc[index]->SetLineColor( kBlack );
+      if( shapeNorm ) {
+        histos_mc[index]->Scale( scaleFactor );
+      } 
+
+      if(i==0) histo_mc = (TH1D*) histos_mc[index]->Clone("histo_mc");
+      else histo_mc->Add(histos_mc[index]);
+      bgStack.Add(histos_mc[index]);
+    }
+
+
+    TCanvas* c1 = new TCanvas("c1", "", 600, 600);
+    c1->cd();
+    
+    TCanvas* c1_log = new TCanvas("c1_log", "", 600, 600);
+
+    float yMaxScale = 1.1;
+    float yMax1 = h1_data->GetMaximum()*yMaxScale;
+    float yMax2 = yMaxScale*(h1_data->GetMaximum() + sqrt(h1_data->GetMaximum()));
+    float yMax3 = yMaxScale*(bgStack.GetMaximum());
+    float yMax = (yMax1>yMax2) ? yMax1 : yMax2;
+    if( yMax3 > yMax ) yMax = yMax3;
+    if( h1_data->GetNbinsX()<2 ) yMax *=3.;
+
+    std::string xAxisTitle;
+    if( units!="" ) 
+      xAxisTitle = (std::string)(Form("%s [%s]", axisName.c_str(), units.c_str()) );
+    else
+      xAxisTitle = (std::string)(Form("%s", axisName.c_str()) );
+
+    std::string yAxisTitle;
+    if(binWidth>0.99){
+      if( units!="" ) 
+	yAxisTitle = (std::string)(Form("Events / (%.0f %s)", binWidth, units.c_str()));
+      else
+	yAxisTitle = (std::string)(Form("Events / (%.0f)", binWidth));
+    }
+    else{
+      if( units!="" ) 
+	yAxisTitle = (std::string)(Form("Events / (%.2f %s)", binWidth, units.c_str()));
+      else
+	yAxisTitle = (std::string)(Form("Events / (%.2f)", binWidth));
+    }
+
+    TH2D* h2_axes = new TH2D("axes", "", 10, xMin, xMax, 10, 0., yMax );
+    h2_axes->SetXTitle(xAxisTitle.c_str());
+    h2_axes->SetYTitle(yAxisTitle.c_str());
+
+    c1->cd();
+    TPad* pad1 = MT2DrawTools::getCanvasMainPad();
+    pad1->Draw();
+    pad1->cd();
+    h2_axes->Draw();
+
+    
+   
+    TH2D* h2_axes_log = new TH2D("axes_log", "", 10, xMin, xMax, 10, 0.1, yMax*2.0 );
+    h2_axes_log->SetXTitle(xAxisTitle.c_str());
+    h2_axes_log->SetYTitle(yAxisTitle.c_str());
+
+    c1_log->cd();
+    TPad* pad1_log = MT2DrawTools::getCanvasMainPad( true );
+    pad1_log->Draw();
+    pad1_log->cd();
+    h2_axes_log->Draw();
+   
+
+
+    std::vector<std::string> niceNames = thisRegion.getNiceNames();
+
+    for( unsigned i=0; i<niceNames.size(); ++i ) {
+      float yMax = 0.9-(float)i*0.05;
+      float yMin = yMax - 0.05;
+      TPaveText* regionText = new TPaveText( 0.18, yMin, 0.55, yMax, "brNDC" );
+      regionText->SetTextSize(0.04);
+      //   regionText->SetTextFont(42);
+      regionText->SetFillColor(0);
+      regionText->SetTextAlign(11);
+      regionText->AddText( niceNames[i].c_str() );
+    }
+    
+
+    TLegend* legend = new TLegend( 0.7, 0.9-(bgYields.size()+1)*0.06, 0.93, 0.9 );
+    legend->SetTextSize(0.04);
+    //  legend->SetTextFont(42);
+    legend->SetFillColor(0);
+    legend->AddEntry( gr_data, "Data", "P" );
+    for( unsigned i=0; i<histos_mc.size(); ++i ) {  
+      legend->AddEntry( histos_mc[i], bgYields[i]->getFullName().c_str(), "F" );
+    }
+
+
+    TPaveText* labelTop = MT2DrawTools::getLabelTop(cfg.lumi());
+    
+    TPaveText* ratioText = new TPaveText( 0.133, -0.051, 0.4, 0.1 , "brNDC" );
+    ratioText->SetTextSize(0.04);
+    //  ratioText->SetTextFont(40);
+    ratioText->SetTextColor(2);
+    ratioText->SetFillColor(0);
+    ratioText->SetTextAlign(11);
+    ratioText->AddText( Form("Data/MC = %.2f +/- %.2f", scaleFactor, error_datamc) );
+    
+    c1->cd();
+    pad1->cd();
+    legend->Draw("same");
+    bgStack.Draw("histo same");
+    gr_data->Draw("p same");
+    labelTop->Draw("same");
+    ratioText->Draw("same");
+
+    gPad->RedrawAxis();
+
+    c1_log->cd();
+    pad1_log->cd();
+    legend->Draw("same");
+    bgStack.Draw("histo same");
+    gr_data->Draw("p same");
+    labelTop->Draw("same");
+    ratioText->Draw("same");
+
+    gPad->RedrawAxis();
+
+    float yMinR=0.0;
+    float yMaxR=2.0;
+
+    TH2D* h2_axes_ratio = MT2DrawTools::getRatioAxes( xMin, xMax, yMinR, yMaxR );
+
+    TGraphAsymmErrors* g_ratio = MT2DrawTools::getRatioGraph(h1_data, histo_mc);
+
+    TLine* line = new TLine(xMin, 1.0, xMax, 1.0);
+    line->SetLineColor(1);
+    
+    TLine* lineSF = MT2DrawTools::getSFLine(integral_data, integral_mc, xMin, xMax);
+    
+    TGraphErrors* SFband = MT2DrawTools::getSFBand(integral_data, error_data, integral_mc, error_mc, xMin, xMax);
+
+
+
+    c1->cd();
+    TPad* pad2 = MT2DrawTools::getCanvasRatioPad();
+    pad2->Draw();
+    pad2->cd();
+
+    h2_axes_ratio->Draw("");
+    line->Draw("same");
+    SFband->Draw("3,same");
+    lineSF->Draw("same");
+    g_ratio->Draw("PE,same");    
+    gPad->RedrawAxis();
+
+
+    c1_log->cd();
+    TPad* pad2_log = MT2DrawTools::getCanvasRatioPad( true );
+    pad2_log->Draw();
+    pad2_log->cd();
+
+    h2_axes_ratio->Draw("");
+    line->Draw("same");
+    SFband->Draw("3,same");
+    lineSF->Draw("same");
+    g_ratio->Draw("PE,same");
+    gPad->RedrawAxis();
+
+
+    c1->SaveAs( Form("%s/%s_%s.eps", fullPathPlots.c_str(), saveName.c_str(), thisRegion.getName().c_str()) );
+    c1->SaveAs( Form("%s/%s_%s.png", fullPathPlots.c_str(), saveName.c_str(), thisRegion.getName().c_str()) );
+    c1->SaveAs( Form("%s/%s_%s.pdf", fullPathPlots.c_str(), saveName.c_str(), thisRegion.getName().c_str()) );
+
+    c1_log->SaveAs( Form("%s/%s_%s_log.eps", fullPathPlots.c_str(), saveName.c_str(), thisRegion.getName().c_str()) );
+    c1_log->SaveAs( Form("%s/%s_%s_log.png", fullPathPlots.c_str(), saveName.c_str(), thisRegion.getName().c_str()) );
+    c1_log->SaveAs( Form("%s/%s_%s_log.pdf", fullPathPlots.c_str(), saveName.c_str(), thisRegion.getName().c_str()) );
+
+    delete c1;
+    delete h2_axes;
+
+    delete c1_log;
+    delete h2_axes_log;
+    
+    delete h2_axes_ratio;
+    
+    delete h1_data;
+  
+    for( unsigned i=0; i<histos_mc.size(); ++i )
+      delete histos_mc[i];
+
+  }// for MT2 regions
+
+}
+
+
+
+
+
+
+
+
+
+float computeScaleFactor( MT2Config cfg, MT2Analysis<MT2EstimateTree>* data, std::vector<MT2Analysis<MT2EstimateTree>* >  bgYields, const std::string& saveName, const std::string& varName, const std::string& selection, int nBins, float xMin, float xMax, std::string axisName, const std::string& units ) {
+
+
+  float binWidth = (xMax-xMin)/nBins;
+  if( axisName=="" ) axisName = varName;
+
+
+  float scaleFactor;
+
+  std::vector<int> colors;
+  if( bgYields.size()==3 ) { // estimates
+    colors.push_back(402); 
+    colors.push_back(430); 
+    colors.push_back(418); 
+  } else { // mc
+    colors.push_back(430); // other=zll
+    colors.push_back(401); // qcd
+    colors.push_back(417); // w+jets
+    colors.push_back(419); // z+jets
+    colors.push_back(855); // top
+    //colors.push_back(); // other
+  }
+
+  std::string fullPathPlots = cfg.getEventYieldDir() + "/plotsDataMC";
+  if( shapeNorm ) fullPathPlots += "_shape";
+  system( Form("mkdir -p %s", fullPathPlots.c_str()) );
+
+  TH1::AddDirectory(kTRUE); // stupid ROOT memory allocation needs this
+
+  std::set<MT2Region> MT2Regions = data->getRegions();
+  
+  for( std::set<MT2Region>::iterator iMT2 = MT2Regions.begin(); iMT2!=MT2Regions.end(); ++iMT2 ) {
+  
+    MT2Region thisRegion( (*iMT2) );
+
+    TTree* tree_data = data->get(thisRegion)->tree;
+    TH1D* h1_data = new TH1D("h1_data", "", nBins, xMin, xMax );
+    tree_data->Project( "h1_data", varName.c_str(), selection.c_str() );
+
+    TGraphAsymmErrors* gr_data = MT2DrawTools::getPoissonGraph(h1_data);
+    gr_data->SetMarkerStyle(20);
+    gr_data->SetMarkerSize(1.2);
+
+
+    std::vector< TH1D* > histos_mc;
+    for( unsigned i=0; i<bgYields.size(); ++i ) { 
+      TTree* tree_mc = (bgYields[i]->get(thisRegion)->tree);
+      std::string thisName = "h1_" + bgYields[i]->getName();
+      TH1D* h1_mc = new TH1D( thisName.c_str(), "", nBins, xMin, xMax );
+      h1_mc->Sumw2();
+      if( selection!="" )
+	//tree_mc->Project( thisName.c_str(), varName.c_str(), Form("%s/puWeight", selection.c_str()) );
+	tree_mc->Project( thisName.c_str(), varName.c_str(), Form("%s", selection.c_str()) );
+      else
+        tree_mc->Project( thisName.c_str(), varName.c_str(), "" );
+
+      histos_mc.push_back(h1_mc);
+    }
+
+    TH1::AddDirectory(kFALSE); // stupid ROOT memory allocation needs this
+
+    TH1D* mc_sum;
+    for( unsigned i=0; i<histos_mc.size(); ++i ) { 
+      if( i==0 ) {
+        mc_sum = new TH1D( *histos_mc[i] );
+        mc_sum->SetName("mc_sum");
+      } else {
+        mc_sum->Add( histos_mc[i] );
+      }
+    }
+
+    std::cout << "Integrals: " << h1_data->Integral(0, nBins+1) << "\t" << mc_sum->Integral(0, nBins+1) << std::endl;
+    scaleFactor = h1_data->Integral(0, nBins+1)/mc_sum->Integral(0, nBins+1);   
+    // if( shapeNorm ) 
+      std::cout << "SF: " << scaleFactor << std::endl;
+ 
+    TH1D* histo_mc;
+    THStack bgStack("bgStack", "");
+    for( unsigned i=0; i<histos_mc.size(); ++i ) { 
+      int index = bgYields.size() - i - 1;
+      histos_mc[index]->SetFillColor( colors[index] );
+      histos_mc[index]->SetLineColor( kBlack );
+      if( shapeNorm ) {
+        histos_mc[index]->Scale( scaleFactor );
+      }
+
+      if(i==0) histo_mc = (TH1D*) histos_mc[index]->Clone("histo_mc");
+      else histo_mc->Add(histos_mc[index]);
+      bgStack.Add(histos_mc[index]);
+    }
+
+
+    TCanvas* c1 = new TCanvas("c1", "", 600, 600);
+    c1->cd();
+    TPad *pad1 = MT2DrawTools::getCanvasMainPad();
+    TPad *pad2 = MT2DrawTools::getCanvasRatioPad();
+        
+    TCanvas* c1_log = new TCanvas("c1_log", "", 600, 600);
+    c1_log->cd();
+    TPad *pad1_log = MT2DrawTools::getCanvasMainPad( true );
+    TPad *pad2_log = MT2DrawTools::getCanvasRatioPad( true );
+ 
+    float yMaxScale = 1.1;
+    float yMax1 = h1_data->GetMaximum()*yMaxScale;
+    float yMax2 = yMaxScale*(h1_data->GetMaximum() + sqrt(h1_data->GetMaximum()));
+    float yMax3 = yMaxScale*(bgStack.GetMaximum());
+    float yMax = (yMax1>yMax2) ? yMax1 : yMax2;
+    if( yMax3 > yMax ) yMax = yMax3;
+    if( h1_data->GetNbinsX()<2 ) yMax *=3.;
+
+    std::string xAxisTitle;
+    if( units!="" ) 
+      xAxisTitle = (std::string)(Form("%s [%s]", axisName.c_str(), units.c_str()) );
+    else
+      xAxisTitle = (std::string)(Form("%s", axisName.c_str()) );
+
+    std::string binWidthText;
+    if( binWidth>=1. )         binWidthText = (std::string)Form("%.0f", binWidth);
+    else if( binWidth>=0.1 )   binWidthText = (std::string)Form("%.1f", binWidth);
+    else if( binWidth>=0.01 )  binWidthText = (std::string)Form("%.2f", binWidth);
+    else if( binWidth>=0.001 ) binWidthText = (std::string)Form("%.3f", binWidth);
+    else                       binWidthText = (std::string)Form("%.4f", binWidth);
+
+   std::string yAxisTitle;
+    if( units!="" ) 
+      yAxisTitle = (std::string)(Form("Events / (%s %s)", binWidthText.c_str(), units.c_str()));
+    else
+      yAxisTitle = (std::string)(Form("Events / (%s)", binWidthText.c_str()));
+
+
+    TH2D* h2_axes = new TH2D("axes", "", 10, xMin, xMax, 10, 0., yMax );
+    h2_axes->SetXTitle(xAxisTitle.c_str());
+    h2_axes->SetYTitle(yAxisTitle.c_str());
+
+    c1->cd();
+    pad1->Draw();
+    pad1->cd();
+    h2_axes->Draw();
+    
+   
+    TH2D* h2_axes_log = new TH2D("axes_log", "", 10, xMin, xMax, 10, 0.1, yMax*2.0 );
+    h2_axes_log->SetXTitle(xAxisTitle.c_str());
+    h2_axes_log->SetYTitle(yAxisTitle.c_str());
+
+    c1_log->cd();
+    pad1_log->Draw();
+    pad1_log->cd();
+    h2_axes_log->Draw();
+   
+
+    std::vector<std::string> niceNames = thisRegion.getNiceNames();
+
+    for( unsigned i=0; i<niceNames.size(); ++i ) {
+      float yMax = 0.9-(float)i*0.05;
+      float yMin = yMax - 0.05;
+      TPaveText* regionText = new TPaveText( 0.18, yMin, 0.55, yMax, "brNDC" );
+      regionText->SetTextSize(0.04);
+      regionText->SetTextFont(42);
+      regionText->SetFillColor(0);
+      regionText->SetTextAlign(11);
+      regionText->AddText( niceNames[i].c_str() );
+
+      pad1->cd();
+      regionText->Draw("same");
+  
+      pad1_log->cd();
+      regionText->Draw("same");
+    }
+    
+    if( shapeNorm ) {
+      TPaveText* normText = new TPaveText( 0.45, 0.8, 0.68, 0.9, "brNDC" );
+      normText->SetFillColor(0);
+      normText->SetTextSize(0.035);
+      normText->AddText( "#splitline{Shape}{Norm.}" );
+      pad1->cd();
+      //normText->Draw("same");
+      pad1_log->cd();
+      // normText->Draw("same");
+    }
+
+    TLegend* legend = new TLegend( 0.7, 0.9-(bgYields.size()+1)*0.06, 0.93, 0.9 );
+    legend->SetTextSize(0.04);
+    legend->SetTextFont(42);
+    legend->SetFillColor(0);
+    legend->AddEntry( gr_data, "Data", "P" );
+    for( unsigned i=0; i<histos_mc.size(); ++i ) {  
+      legend->AddEntry( histos_mc[i], bgYields[i]->getFullName().c_str(), "F" );
+    }
+
+
+    TPaveText* labelTop = MT2DrawTools::getLabelTop(cfg.lumi());
+    
+    TPaveText* ratioText = new TPaveText( 0.133, -0.051, 0.4, 0.1 , "brNDC" );
+    ratioText->SetTextSize(0.04);
+    ratioText->SetTextFont(40);
+    ratioText->SetTextColor(2);
+    ratioText->SetFillColor(0);
+    ratioText->SetTextAlign(11);
+    ratioText->AddText( Form("Data/MC = %.2f", scaleFactor) );
+    //  ratioText->AddText( Form("Data/MC = %.2f +/- %.2f", scaleFactor, error_datamc) );
+     
+
+    TLine* line = new TLine(xMin, 1.0, xMax, 1.0);
+    line->SetLineColor(1);
+    
+    TLine* lineSF = new TLine(xMin, scaleFactor, xMax, scaleFactor);
+    lineSF->SetLineColor(2);
+
+    float yMinR=0.0;
+    float yMaxR=2.0;
+
+  
+    TH2D* h2_axes_ratio = MT2DrawTools::getRatioAxes( xMin, xMax, yMinR, yMaxR );
+    TGraphAsymmErrors* g_ratio = MT2DrawTools::getRatioGraph(h1_data, histo_mc);
+ 
+    TLine* lineCentral = new TLine(xMin, 1.0, xMax, 1.0);
+    lineCentral->SetLineColor(1);
+    TGraphErrors* systBand = MT2DrawTools::getSystBand(xMin, xMax, lumiErr);
+   
+    //    TH1D* mcBand = MT2DrawTools::getMCBandHisto( histo_mc, lumiErr );
+    TF1* fSF = MT2DrawTools::getSFFit(g_ratio, xMin, xMax);
+    TGraphErrors* SFFitBand = MT2DrawTools::getSFFitBand(fSF, xMin, xMax);
+    TPaveText* fitText = MT2DrawTools::getFitText( fSF );
+
+
+    c1->cd();
+    pad1->cd();
+    legend->Draw("same");
+    bgStack.Draw("histo same");
+    gr_data->Draw("p same");
+    labelTop->Draw("same");
+    if( !shapeNorm )
+      fitText->Draw("same");
+    // ratioText->Draw("same");
+  
+    gPad->RedrawAxis();
+
+    c1_log->cd();
+    pad1_log->cd();
+    legend->Draw("same");
+    bgStack.Draw("histo same");
+    gr_data->Draw("p same");
+    labelTop->Draw("same");
+    if( !shapeNorm )
+     fitText->Draw("same");
+    //  ratioText->Draw("same");
+
+    gPad->RedrawAxis();
+
+   /*
+    TLine* line = new TLine(xMin, 1.0, xMax, 1.0);
+    line->SetLineColor(1);
+    
+    TLine* lineSF = MT2DrawTools::getSFLine(integral_data, integral_mc, xMin, xMax);
+   
+    TGraphErrors* SFband = MT2DrawTools::getSFBand(integral_data, error_data, integral_mc, error_mc, xMin, xMax);
+    */
+
+
+    c1->cd();
+    //   TPad* pad2 = MT2DrawTools::getCanvasRatioPad();
+    pad2->Draw();
+    pad2->cd();
+
+    h2_axes_ratio->Draw("");
+ 
+    /*  line->Draw("same");
+    SFband->Draw("3,same");
+    lineSF->Draw("same");
+    */
+    lineCentral->Draw("same");
+    if( !shapeNorm ){
+
+      systBand->Draw("3,same");
+      lineCentral->Draw("same");
+
+      SFFitBand->Draw("3,same");
+      fSF->Draw("same");
+    }
+
+    g_ratio->Draw("PE,same");    
+    gPad->RedrawAxis();
+
+
+    c1_log->cd();
+    pad2_log->Draw();
+    pad2_log->cd();
+
+    h2_axes_ratio->Draw(""); 
+    
+    lineCentral->Draw("same");
+    if( !shapeNorm ){
+
+      systBand->Draw("3,same");
+      lineCentral->Draw("same");
+
+      SFFitBand->Draw("3,same");
+      fSF->Draw("same");
+    }
+    /*
+    line->Draw("same");
+    SFband->Draw("3,same");
+    lineSF->Draw("same"); */
+    g_ratio->Draw("PE,same");
+    gPad->RedrawAxis();
+
+
+    c1->SaveAs( Form("%s/%s_%s.eps", fullPathPlots.c_str(), saveName.c_str(), thisRegion.getName().c_str()) );
+    c1->SaveAs( Form("%s/%s_%s.png", fullPathPlots.c_str(), saveName.c_str(), thisRegion.getName().c_str()) );
+    c1->SaveAs( Form("%s/%s_%s.pdf", fullPathPlots.c_str(), saveName.c_str(), thisRegion.getName().c_str()) );
+
+    c1_log->SaveAs( Form("%s/%s_%s_log.eps", fullPathPlots.c_str(), saveName.c_str(), thisRegion.getName().c_str()) );
+    c1_log->SaveAs( Form("%s/%s_%s_log.png", fullPathPlots.c_str(), saveName.c_str(), thisRegion.getName().c_str()) );
+    c1_log->SaveAs( Form("%s/%s_%s_log.pdf", fullPathPlots.c_str(), saveName.c_str(), thisRegion.getName().c_str()) );
+
+    delete c1;
+    delete h2_axes;
+
+    delete c1_log;
+    delete h2_axes_log;
+    
+    delete h2_axes_ratio;
+    
+    delete h1_data;
+  
+    for( unsigned i=0; i<histos_mc.size(); ++i )
+      delete histos_mc[i];
+
+  }// for MT2 regions
+
+  return scaleFactor;
+
+}
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+void drawYieldsScaled( MT2Config cfg, MT2Analysis<MT2EstimateTree>* data, std::vector<MT2Analysis<MT2EstimateTree>* >  bgYields, const std::string& saveName, const std::string& varName, const std::string& selection, int nBins, float xMin, float xMax, std::string axisName, const std::string& units, float scale ) {
+
+
+  float binWidth = (xMax-xMin)/nBins;
+  if( axisName=="" ) axisName = varName;
+
+
+
+  std::vector<int> colors;
+  if( bgYields.size()==3 ) { // estimates
+    colors.push_back(402); 
+    colors.push_back(430); 
+    colors.push_back(418); 
+  } else { // mc
+    colors.push_back(430); // other=zll
+    colors.push_back(401); // qcd
+    colors.push_back(417); // w+jets
+    colors.push_back(419); // z+jets
+    colors.push_back(855); // top
+    //colors.push_back(); // other
+  }
+
+  std::string fullPathPlots = cfg.getEventYieldDir() + "/plotsDataMC";
+  system( Form("mkdir -p %s", fullPathPlots.c_str()) );
+
+  TH1::AddDirectory(kTRUE); // stupid ROOT memory allocation needs this
+
+  std::set<MT2Region> MT2Regions = data->getRegions();
+  
+  for( std::set<MT2Region>::iterator iMT2 = MT2Regions.begin(); iMT2!=MT2Regions.end(); ++iMT2 ) {
+  
+    MT2Region thisRegion( (*iMT2) );
+
+    TTree* tree_data = data->get(thisRegion)->tree;
+    TH1D* h1_data = new TH1D("h1_data", "", nBins, xMin, xMax );
+    tree_data->Project( "h1_data", varName.c_str(), selection.c_str() );
+
+    TGraphAsymmErrors* gr_data = MT2DrawTools::getPoissonGraph(h1_data);
+    gr_data->SetMarkerStyle(20);
+    gr_data->SetMarkerSize(1.2);
+
+
+    std::vector< TH1D* > histos_mc;
+    for( unsigned i=0; i<bgYields.size(); ++i ) { 
+      TTree* tree_mc = (bgYields[i]->get(thisRegion)->tree);
+      std::string thisName = "h1_" + bgYields[i]->getName();
+      TH1D* h1_mc = new TH1D( thisName.c_str(), "", nBins, xMin, xMax );
+      h1_mc->Sumw2();
+      if( selection!="" )
+	tree_mc->Project( thisName.c_str(), varName.c_str(), Form("%s", selection.c_str()) );
+      else
+        tree_mc->Project( thisName.c_str(), varName.c_str(), "" );
+
+      if(i==4) h1_mc->Scale(scale);
+
+      histos_mc.push_back(h1_mc);
+    }
+
+    TH1::AddDirectory(kFALSE); // stupid ROOT memory allocation needs this
+
+    TH1D* mc_sum;
+    for( unsigned i=0; i<histos_mc.size(); ++i ) { 
+      if( i==0 ) {
+        mc_sum = new TH1D( *histos_mc[i] );
+        mc_sum->SetName("mc_sum");
+      } else {
+        mc_sum->Add( histos_mc[i] );
+      }
+    }
+
+    std::cout << "Integrals: " << h1_data->Integral(0, nBins+1) << "\t" << mc_sum->Integral(0, nBins+1) << std::endl;
+    float scaleFactor = h1_data->Integral(0, nBins+1)/mc_sum->Integral(0, nBins+1);   
+    if( shapeNorm ) 
+      std::cout << "SF: " << scaleFactor << std::endl;
+
+    double error_data;
+    double integral_data = h1_data->IntegralAndError(0, nBins+1, error_data);
+
+    double error_mc;
+    double integral_mc = mc_sum->IntegralAndError(0, nBins+1, error_mc);
+
+    double error_datamc = MT2DrawTools::getSFError(integral_data, error_data, integral_mc, error_mc);
+
+    TH1D* histo_mc;
+    THStack bgStack("bgStack", "");
+    for( unsigned i=0; i<histos_mc.size(); ++i ) { 
+      int index = bgYields.size() - i - 1;
+      histos_mc[index]->SetFillColor( colors[index] );
+      histos_mc[index]->SetLineColor( kBlack );
+      if( shapeNorm ) {
+        histos_mc[index]->Scale( scaleFactor );
+      } 
+
+      if(i==0) histo_mc = (TH1D*) histos_mc[index]->Clone("histo_mc");
+      else histo_mc->Add(histos_mc[index]);
+      bgStack.Add(histos_mc[index]);
+    }
+
+
+    TCanvas* c1 = new TCanvas("c1", "", 600, 600);
+    c1->cd();
+    
+    TCanvas* c1_log = new TCanvas("c1_log", "", 600, 600);
+
+    float yMaxScale = 1.1;
+    float yMax1 = h1_data->GetMaximum()*yMaxScale;
+    float yMax2 = yMaxScale*(h1_data->GetMaximum() + sqrt(h1_data->GetMaximum()));
+    float yMax3 = yMaxScale*(bgStack.GetMaximum());
+    float yMax = (yMax1>yMax2) ? yMax1 : yMax2;
+    if( yMax3 > yMax ) yMax = yMax3;
+    if( h1_data->GetNbinsX()<2 ) yMax *=3.;
+
+    std::string xAxisTitle;
+    if( units!="" ) 
+      xAxisTitle = (std::string)(Form("%s [%s]", axisName.c_str(), units.c_str()) );
+    else
+      xAxisTitle = (std::string)(Form("%s", axisName.c_str()) );
+
+    std::string yAxisTitle;
+    if(binWidth>0.99){
+      if( units!="" ) 
+	yAxisTitle = (std::string)(Form("Events / (%.0f %s)", binWidth, units.c_str()));
+      else
+	yAxisTitle = (std::string)(Form("Events / (%.0f)", binWidth));
+    }
+    else{
+      if( units!="" ) 
+	yAxisTitle = (std::string)(Form("Events / (%.2f %s)", binWidth, units.c_str()));
+      else
+	yAxisTitle = (std::string)(Form("Events / (%.2f)", binWidth));
+    }
+
+    TH2D* h2_axes = new TH2D("axes", "", 10, xMin, xMax, 10, 0., yMax );
+    h2_axes->SetXTitle(xAxisTitle.c_str());
+    h2_axes->SetYTitle(yAxisTitle.c_str());
+
+    c1->cd();
+    TPad* pad1 = MT2DrawTools::getCanvasMainPad();
+    pad1->Draw();
+    pad1->cd();
+    h2_axes->Draw();
+
+    
+   
+    TH2D* h2_axes_log = new TH2D("axes_log", "", 10, xMin, xMax, 10, 0.1, yMax*2.0 );
+    h2_axes_log->SetXTitle(xAxisTitle.c_str());
+    h2_axes_log->SetYTitle(yAxisTitle.c_str());
+
+    c1_log->cd();
+    TPad* pad1_log = MT2DrawTools::getCanvasMainPad( true );
+    pad1_log->Draw();
+    pad1_log->cd();
+    h2_axes_log->Draw();
+   
+
+
+    std::vector<std::string> niceNames = thisRegion.getNiceNames();
+
+    for( unsigned i=0; i<niceNames.size(); ++i ) {
+      float yMax = 0.9-(float)i*0.05;
+      float yMin = yMax - 0.05;
+      TPaveText* regionText = new TPaveText( 0.18, yMin, 0.55, yMax, "brNDC" );
+      regionText->SetTextSize(0.04);
+      //   regionText->SetTextFont(42);
+      regionText->SetFillColor(0);
+      regionText->SetTextAlign(11);
+      regionText->AddText( niceNames[i].c_str() );
+    }
+    
+
+    TLegend* legend = new TLegend( 0.7, 0.9-(bgYields.size()+1)*0.06, 0.93, 0.9 );
+    legend->SetTextSize(0.04);
+    //  legend->SetTextFont(42);
+    legend->SetFillColor(0);
+    legend->AddEntry( gr_data, "Data", "P" );
+    for( unsigned i=0; i<histos_mc.size(); ++i ) {  
+      legend->AddEntry( histos_mc[i], bgYields[i]->getFullName().c_str(), "F" );
+    }
+
+
+    TPaveText* labelTop = MT2DrawTools::getLabelTop(cfg.lumi());
+    
+    TPaveText* ratioText = new TPaveText( 0.133, -0.051, 0.4, 0.1 , "brNDC" );
+    ratioText->SetTextSize(0.04);
+    //  ratioText->SetTextFont(40);
+    ratioText->SetTextColor(2);
+    ratioText->SetFillColor(0);
+    ratioText->SetTextAlign(11);
+    ratioText->AddText( Form("Data/MC = %.2f +/- %.2f", scaleFactor, error_datamc) );
+    
+    c1->cd();
+    pad1->cd();
+    legend->Draw("same");
+    bgStack.Draw("histo same");
+    gr_data->Draw("p same");
+    labelTop->Draw("same");
+    ratioText->Draw("same");
+
+    gPad->RedrawAxis();
+
+    c1_log->cd();
+    pad1_log->cd();
+    legend->Draw("same");
+    bgStack.Draw("histo same");
+    gr_data->Draw("p same");
+    labelTop->Draw("same");
+    ratioText->Draw("same");
+
+    gPad->RedrawAxis();
+
+    float yMinR=0.0;
+    float yMaxR=2.0;
+
+    TH2D* h2_axes_ratio = MT2DrawTools::getRatioAxes( xMin, xMax, yMinR, yMaxR );
+    TGraphAsymmErrors* g_ratio = MT2DrawTools::getRatioGraph(h1_data, histo_mc);
+    TLine* line = new TLine(xMin, 1.0, xMax, 1.0);
+    line->SetLineColor(1);
+    TLine* lineSF = MT2DrawTools::getSFLine(integral_data, integral_mc, xMin, xMax);
+    TGraphErrors* SFband = MT2DrawTools::getSFBand(integral_data, error_data, integral_mc, error_mc, xMin, xMax);
+
+    c1->cd();
+    TPad* pad2 = MT2DrawTools::getCanvasRatioPad();
+    pad2->Draw();
+    pad2->cd();
+
+    h2_axes_ratio->Draw("");
+    line->Draw("same");
+    SFband->Draw("3,same");
+    lineSF->Draw("same");
+    g_ratio->Draw("PE,same");    
+    gPad->RedrawAxis();
+
+    c1_log->cd();
+    TPad* pad2_log = MT2DrawTools::getCanvasRatioPad( true );
+    pad2_log->Draw();
+    pad2_log->cd();
+
+    h2_axes_ratio->Draw("");
+    line->Draw("same");
+    SFband->Draw("3,same");
+    lineSF->Draw("same");
+    g_ratio->Draw("PE,same");
+    gPad->RedrawAxis();
+
+    c1->SaveAs( Form("%s/%s_%s.eps", fullPathPlots.c_str(), saveName.c_str(), thisRegion.getName().c_str()) );
+    c1->SaveAs( Form("%s/%s_%s.png", fullPathPlots.c_str(), saveName.c_str(), thisRegion.getName().c_str()) );
+    c1->SaveAs( Form("%s/%s_%s.pdf", fullPathPlots.c_str(), saveName.c_str(), thisRegion.getName().c_str()) );
+
+    c1_log->SaveAs( Form("%s/%s_%s_log.eps", fullPathPlots.c_str(), saveName.c_str(), thisRegion.getName().c_str()) );
+    c1_log->SaveAs( Form("%s/%s_%s_log.png", fullPathPlots.c_str(), saveName.c_str(), thisRegion.getName().c_str()) );
+    c1_log->SaveAs( Form("%s/%s_%s_log.pdf", fullPathPlots.c_str(), saveName.c_str(), thisRegion.getName().c_str()) );
+
+    delete c1;
+    delete h2_axes;
+
+    delete c1_log;
+    delete h2_axes_log;
+    
+    delete h2_axes_ratio;
+    
+    delete h1_data;
+  
+    for( unsigned i=0; i<histos_mc.size(); ++i )
+      delete histos_mc[i];
+
+  }// for MT2 regions
+
+}
+
